requirement.txt
pandas==2.2.2
numpy==1.25.2
matplotlib==3.7.3
seaborn==0.12.2
scikit-learn==1.5.0
xgboost==1.7.4
catboost==1.2.5
sqlalchemy==2.0.19
pymysql==1.0.3
joblib==1.3.2
category_encoders==2.6.0
gunicorn==20.1.0
waitress==2.1.2
flask==3.0.0
ipykernel==6.23.1
requests==2.31.0
python-dotenv==1.0.1